langgraph_server/requirements.txt
langchain
langchain-core
langchain-openai
langgraph
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
langgraph-cli[inmem]
